Added support for a messageAttr for each error type
It is now possible to define multiple data-messages with each one
handling a different constraint violation. As an example

  <input type="email" data-message-type="This is not an email"
         data-message-required="This field is required"
         required />

... defines two constraint violation messages. One for a violation
related to the type of the input field (like email in this example)
and one for the required-attribute.

The used attribute is determined by the conf.messageAttrTypeSuffixes
object in combination with conf.messageAttr.