Moved cached tooltip instance into non-clashing namespace
The tooltip module allows for the actual tooltip container to be
specified using a data-tooltip attribute at the trigger. This
caused a clash with the cached tooltip-object instance which
was stored in trigger's data store under the name "tooltip".

This commit changes the name of the cached tooltip to
"jquerytools.tooltip".